Fixed admin dashboard
